recreate all files since 1st batch.
@teixeirak, I moved what we already sent into a subfolder called "Canceled 1st and 2nd batch" (in 2-approved/processed and in 4-transferred-to_EFDB)

I am leaving everything in 1-to-review, I'll let you move them to approve. But notice that I also left the ones you already approved in 2-aproved, so you may want to delete a file from 1 if you already fixed it by hand and moved it to 2. Or you can overwrite/delete the ones in 2 if needed. Crowell should stay in 2 as it is note created anymore (D.precedence reverted)

Let me know if you have questions