Pushing corrected version of Wirth_2002_fast
@teixeirak You had hand edited this citation here: https://github.com/forc-db/IPCC-EFDB-integration/commit/9511c7445454c8c97bc7923d4f481e8077ee9db8

You may have to do so again before approving. I need to get my head back into this so that you don't need to hand edit anymore.... maybe when you get a chance you can remind me of what needs to happen. I have in my head that it is quite an involved process but maybe it isn't and could fix it quickly.

@mawilliams99, tagging you here, just FYI